doc:HTTPFuzzerEditorMenu文件中的常量位置修改并修改引用
