+++ app/renderer/src/main/src/defaultConstants/HTTPFuzzerPage.ts
@@ -3,6 +3,7 @@
  * @author luoluo
  */
 
+import {LabelDataProps} from "@/pages/fuzzer/HTTPFuzzerEditorMenu"
 import {AdvancedConfigShowProps, FuzzerResponse} from "@/pages/fuzzer/HTTPFuzzerPage"
 import {AdvancedConfigValueProps} from "@/pages/fuzzer/HttpQueryAdvancedConfig/HttpQueryAdvancedConfigType"
 import {WebFuzzerPageInfoProps} from "@/store/pageInfo"
@@ -153,3 +154,53 @@ export const defaultWebFuzzerPageInfo: WebFuzzerPageInfoProps = {
     request: defaultPostTemplate,
     variableActiveKeys: undefined
 }
+// 注：此处顺序为倒序（新增DefaultDescription记得带-fixed，此处为标识固定项）
+export const defaultLabel: LabelDataProps[] = [
+    {
+        DefaultDescription: "反向正则（单个）-fixed",
+        Description: "反向正则（单个）",
+        Label: "{{regen:one([0-9a-f]{3})}}"
+    },
+    {
+        DefaultDescription: "反向正则（全部）-fixed",
+        Description: "反向正则（全部）",
+        Label: "{{regen([0-9a-f]{3})}}"
+    },
+    {
+        DefaultDescription: "时间戳（秒）-fixed",
+        Description: "时间戳（秒）",
+        Label: "{{timestamp(seconds)}}"
+    },
+    {
+        DefaultDescription: "验证码-fixed",
+        Description: "验证码",
+        Label: "{{int(0000-9999)}}"
+    },
+    {
+        DefaultDescription: "随机数-fixed",
+        Description: "随机数",
+        Label: "{{randint(0,10)}}"
+    },
+    {
+        DefaultDescription: "随机字符串-fixed",
+        Description: "随机字符串",
+        Label: "{{randstr}}"
+    },
+    {
+        DefaultDescription: "整数范围-fixed",
+        Description: "整数范围",
+        Label: "{{int(1-10)}}"
+    },
+    {
+        DefaultDescription: "插入Payload-fixed",
+        Description: "插入Payload"
+    },
+    {
+        DefaultDescription: "插入临时字典-fixed",
+        Description: "插入临时字典"
+    },
+    {
+        DefaultDescription: "插入文件-fixed",
+        Description: "插入文件"
+    }
+]
